requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
asttokens==2.4.0
autopep8==1.7.0
backcall==0.2.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dj-database-url==3.1.0
Django==5.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.128.0
Flask==3.1.2
fonttools==4.43.1
gunicorn==24.1.1
h11==0.16.0
idna==3.11
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mlxtend==0.23.0
nest-asyncio==1.5.8
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.5.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycodestyle==2.9.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.16.1
pyparsing==3.1.1
pytest==8.2.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.32.5
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.50.0
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.11.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
urllib3==2.5.0
uvicorn==0.40.0
wcwidth==0.2.8
Werkzeug==3.1.3
whitenoise==6.11.0
xlrd==2.0.1
